streamlit/requirements.txt
streamlit>=1.30.0
clickhouse-connect
pandas
python-dotenv
